requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
boto3==1.18.42
botocore==1.21.42
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.3
Django==3.2.7
django-admin-list-filter-dropdown==1.0.3
django-admin-multiple-choice-list-filter==0.1.1
django-admin-rangefilter==0.8.3
django-allauth==0.48.0
django-more-admin-filters==1.3
django-oauth-toolkit==1.7.0
django-phonenumber-field==6.0.0
django-storages==1.12.1
djangorestframework==3.12.4
djangorestframework-jwt
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
drf-yasg2==1.19.4
fcm-django==1.0.5
firebase-admin==5.0.2
google-api-core==1.31.2
google-api-python-client==2.21.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==2.0.0
google-cloud-firestore==2.3.2
google-cloud-storage==1.42.1
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.40.0
httplib2==0.19.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jwcrypto==1.0
MarkupSafe==2.0.1
msgpack==1.0.2
# mysqlclient==2.0.3
mysqlclient==1.4.6
oauthlib==3.2.0
packaging==21.0
phonenumberslite==8.12.43
Pillow==8.3.2
proto-plus==1.19.0
protobuf==3.17.3
PyArabic==0.6.14
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==1.7.1
pyotp==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
social-auth-core
sqlparse==0.4.2
Tashaphyne==0.3.4.1
typing-extensions==3.10.0.2
unicode==2.8
Unidecode==1.3.2
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.13.3